Added transaction log

+++ atm_app/forms.py
@@ -8,7 +8,7 @@ currency_choices = (("USD", "USD"),
                     ("GBP", "GBP"))
 
 class AmountForm(forms.Form):
-	amount = forms.IntegerField(label="Amount", widget=forms.TextInput(attrs={'style' : 'width: 300px;', 'class': 'form-control'}))
+	amount = forms.DecimalField(label="Amount", widget=forms.TextInput(attrs={'style' : 'width: 300px;', 'class': 'form-control'}))
 	currency = forms.ChoiceField(label="Currency", widget=forms.Select, choices=currency_choices)
 
 class NewUserForm(UserCreationForm):
@@ -18,10 +18,10 @@ class NewUserForm(UserCreationForm):
 
 class ExchangeForm(forms.Form):
         from_currency = forms.ChoiceField(label="From Currency", choices=currency_choices, widget=forms.Select)
-        amount = forms.IntegerField(label="Amount")
+        amount = forms.DecimalField(label="Amount")
         to_currency = forms.ChoiceField(label="To Currency", choices=currency_choices, widget=forms.Select)
 
 class TransferForm(forms.Form):
         recipient_user = forms.CharField(label="Recipient User")
-        amount = forms.IntegerField(label="Amount")
+        amount = forms.DecimalField(label="Amount")
         currency = forms.ChoiceField(label="Currency", choices=currency_choices, widget=forms.Select)
